mnist2_api/requirements_gpu.txt
flask
flask_restful
opencv-python
pillow
tensorflow-gpu
Keras